LPS-69093 Revert Journal related ExportImportContentProcessor changes
