+++ modules/apps/web-experience/journal/journal-service/build.gradle
@@ -9,8 +9,8 @@ dependencies {
 	provided group: "biz.aQute.bnd", name: "biz.aQute.bndlib", version: "3.1.0"
 	provided group: "com.liferay", name: "com.liferay.dynamic.data.mapping.api", version: "3.0.0"
 	provided group: "com.liferay", name: "com.liferay.dynamic.data.mapping.service", version: "2.0.0"
-	provided group: "com.liferay", name: "com.liferay.exportimport.api", version: "3.0.0-20160907.160400-1"
-	provided group: "com.liferay", name: "com.liferay.exportimport.service", version: "4.0.0-20160913.171549-2"
+	provided group: "com.liferay", name: "com.liferay.exportimport.api", version: "2.0.0"
+	provided group: "com.liferay", name: "com.liferay.exportimport.service", version: "3.3.0"
 	provided group: "com.liferay", name: "com.liferay.mobile.device.rules.api", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.osgi.service.tracker.collections", version: "2.0.0"
 	provided group: "com.liferay", name: "com.liferay.osgi.util", version: "3.0.0"
